WIP
This was an initial stab at switching this repository over to use
Chocolatey.Cake.Recipe, as well as to switch to using the new csproj
format (to allow compiling for .NET Standard 2.0).  The thought process
here was that since this is a library, and we are targetting at a
monimum going forward .NET 4.8, that is would make sense to set the
project up to be more easily cross-compiled to different target
frameworks when required.

The reason that this work was halted was due to the fact that the
current tests require an API key in order to run, and as a result, they
run into problems with a key that only allows 4 requests per hour.

For now, going to halt work on this, until we decide how we want to
proceed, i.e. either to re-write the tests to _not_ rely on an API Key,
or use a different key that doesn't have the same limitations.